requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
# Flask-Mail==0.9.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-decouple==3.4
python-dotenv==0.19.2
SQLAlchemy==1.4.27
Werkzeug==2.0.2
WTForms==3.0.0
python-decouple==3.4
email_validator==1.1.3
wheel==0.33.1
GitPython==3.1.31
PyGithub==1.58.0
numpy==1.24.2
scikit-learn==0.24.2